requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
gunicorn==21.2.0
SQLAlchemy==2.0.34
alembic==1.13.2
psycopg2-binary==2.9.9
pydantic==2.7.4
email-validator==2.2.0
python-dotenv==1.0.1
argon2-cffi==23.1.0
PyJWT==2.9.0
httpx==0.27.2
prometheus-fastapi-instrumentator==6.1.0
structlog==24.1.0
slowapi==0.1.9
